delete original pretraining.py
